+++ src/domain/repositories/log.repository.ts
@@ -1,9 +1,9 @@
 import type { LogEntity } from "../entities/log.entity.js";
-import type { LogSeverity } from "../enums/logSeverity.enum.js";
-import type { DeleteLogsOptions } from "../interfaces/deleteLogsOptions.interface.js";
+import type { LogSeverity } from "../types/enums/logSeverity.enum.js";
+import type { FilterLogsOptions } from "../types/interfaces/filterLogsOptions.interface.js";
 
 export abstract class LogRepository {
-    abstract readLogs(severity?: LogSeverity): Promise<LogEntity[]>;
+    abstract readLogs(options?: FilterLogsOptions): Promise<LogEntity[]>;
     abstract saveLog(log: LogEntity): Promise<void>;
-    abstract deleteLogs(options: DeleteLogsOptions): Promise<void>;
+    abstract deleteLogs(options: FilterLogsOptions): Promise<void>;
 };