+++ src/index.ts
@@ -1,2 +1,2 @@
-export { LogSeverity } from "./domain/enums/logSeverity.enum.js";
+export { LogSeverity } from "./domain/types/enums/logSeverity.enum.js";
 export { createLogger } from "./createLogger.js";